% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{coverMatrix}
\alias{coverMatrix}
\alias{coverMatrix.Rcpp_POSet}
\title{Poset cover matrix}
\usage{
coverMatrix(x)
}
\arguments{
\item{x}{an object of class \code{\link[POSetR:poset]{Rcpp_POSet}}.}
}
\value{
a logical square matrix.
}
\description{
The function returns the cover matrix for objects of class \code{\link[POSetR:poset]{Rcpp_POSet}}.
}
\examples{
dom <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)
p <- poset(x = dom)
coverMatrix(p)
}
