% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{incidenceMatrix}
\alias{incidenceMatrix}
\alias{incidenceMatrix.poset}
\title{Poset incidence matrix}
\usage{
incidenceMatrix(x)
}
\arguments{
\item{x}{an object of class \code{\link[POSetR:poset]{poset}}.}
}
\value{
a logical square matrix.
}
\description{
The function returns the incidence matrix for objects of class \code{\link[POSetR:poset]{poset}}.
}
\examples{
dom <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)
p <- poset(x = dom)
incidenceMatrix(p)
}
