% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Rcpp_POSet.R
\name{summary.Rcpp_POSet}
\alias{summary.Rcpp_POSet}
\title{Poset summary}
\usage{
\method{summary}{Rcpp_POSet}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link[POSetR:poset]{Rcpp_POSet}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An S3 object of class \code{summary_poset} listing and counting the poset elements, their strict dominances and their incomparabilities.
}
\description{
Method of the \code{\link[base]{summary}} function for objects of class \code{\link[POSetR:poset]{Rcpp_POSet}}.
}
\examples{
dom <- matrix(c(
  "a", "b",
  "c", "b",
  "b", "d"
), ncol = 2, byrow = TRUE)
p <- poset(x = dom)
summary(p)

summary(chain(1:4))
summary(antiChain(LETTERS[1:5]))
}
